requirements.txt
Flask
AfricastalkingGateway
